requirements.txt
pygments==2.0.1
django==1.7
djangorestframework==3.0.4
